Module seccompiler::compiler

Module defining the logic for compiling the deserialized filter objects into the IR. Used by seccompiler-bin.

Via the Compiler::compile_blob() method, it also drives the entire JSON -> BLOB transformation process.

It also defines some of the objects that a JSON seccomp filter is deserialized into: Filter, SyscallRule. The rest of objects are deserialized directly into the IR (intermediate representation): SeccompCondition, SeccompAction, SeccompCmpOp, SeccompCmpArgLen.

Structs

Compiler

Object responsible for compiling Filters into BpfPrograms. Uses the SeccompFilter interface as an IR language.

Filter

Deserializable seccomp filter. Refers to one thread category.

JsonFile

Deserializable object that represents the Json filter file.

SyscallRule

Deserializable object representing a syscall rule.

Enums

CompilationError

Errors compiling Filters into BPF.

Constants

_DERIVE_Display_FOR_CompilationError 🔒